Fix invalid array access bug in readable sketch
Readable sketch would panic if the target idx was past
the end of the data array.

This occurs when the item count is greater than the sum
of weights in the sketch.  In KLL sketches, this can happen when
a compactor has an odd capacity; since the compactor doubles each item's weight,
the total output weight will never exactly equal the total input weight.

To be safe, I've also included the sampler's stored weighted value
in the readable sketch.  This eliminates one other source of inaccuracy for quantile queries.