Fix daemon recovery from DNS change
Previous behavior was to resolve the daemon's publish address to an IP address on startup.  This prevents the daemon from recovering if the DNS entry changes to point to a new IP address.

The fix is to defer resolving the DNS address to an IP address until the daemon connects to the server.  That way, if the DNS address changes, the client will resolve the address again when it attempts to reconnect.

I've also set the default window size of the daemon to 10 to make it slightly faster to test this kind of thing manually.